refactor(blocks): remove a bunch of custom slabs and stairs
