+++ src/main/java/com/chimericdream/minekea/block/building/general/CrackedBasaltBricksBlock.java
@@ -1,4 +1,4 @@
-package com.chimericdream.minekea.block.building.basalt_bricks;
+package com.chimericdream.minekea.block.building.general;
 
 import com.chimericdream.minekea.ModInfo;
 import com.chimericdream.minekea.resource.LootTable;
@@ -11,7 +11,6 @@ import net.devtech.arrp.json.blockstate.JState;
 import net.devtech.arrp.json.models.JModel;
 import net.devtech.arrp.json.models.JTextures;
 import net.devtech.arrp.json.recipe.JIngredient;
-import net.devtech.arrp.json.recipe.JIngredients;
 import net.devtech.arrp.json.recipe.JRecipe;
 import net.devtech.arrp.json.recipe.JResult;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
@@ -23,10 +22,10 @@ import net.minecraft.item.ItemGroup;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 
-public class MossyBasaltBricksBlock extends Block implements MinekeaBlock {
-    public static final Identifier BLOCK_ID = new Identifier(ModInfo.MOD_ID, "building/mossy_basalt_bricks");
+public class CrackedBasaltBricksBlock extends Block implements MinekeaBlock {
+    public static final Identifier BLOCK_ID = new Identifier(ModInfo.MOD_ID, "building/cracked_basalt_bricks");
 
-    public MossyBasaltBricksBlock() {
+    public CrackedBasaltBricksBlock() {
         super(FabricBlockSettings.copyOf(Blocks.SMOOTH_BASALT));
     }
 
@@ -50,11 +49,16 @@ public class MossyBasaltBricksBlock extends Block implements MinekeaBlock {
 
         MinekeaResourcePack.RESOURCE_PACK.addRecipe(
             BLOCK_ID,
-            JRecipe.shapeless(
-                JIngredients.ingredients()
-                    .add(JIngredient.ingredient().item(BasaltBricksBlock.BLOCK_ID.toString()))
-                    .add(JIngredient.ingredient().item("minecraft:vine")),
+            JRecipe.smelting(
+                JIngredient.ingredient().item(BasaltBricksBlock.BLOCK_ID.toString()),
                 JResult.result(BLOCK_ID.toString())
+            )
+        );
+        MinekeaResourcePack.RESOURCE_PACK.addRecipe(
+            new Identifier(BLOCK_ID.getNamespace(), BLOCK_ID.getPath() + "_stonecutting"),
+            JRecipe.stonecutting(
+                JIngredient.ingredient().item(BasaltBricksBlock.BLOCK_ID.toString()),
+                JResult.stackedResult(BLOCK_ID.toString(), 1)
             )
         );
         MinekeaResourcePack.RESOURCE_PACK.addRecipe(
