+++ src/main/java/com/chimericdream/minekea/block/building/general/WarpedNetherBricksBlock.java
@@ -1,6 +1,7 @@
-package com.chimericdream.minekea.block.building.basalt_bricks;
+package com.chimericdream.minekea.block.building.general;
 
 import com.chimericdream.minekea.ModInfo;
+import com.chimericdream.minekea.crops.WarpedWartItem;
 import com.chimericdream.minekea.resource.LootTable;
 import com.chimericdream.minekea.resource.MinekeaResourcePack;
 import com.chimericdream.minekea.resource.Model;
@@ -10,25 +11,21 @@ import net.devtech.arrp.json.blockstate.JBlockModel;
 import net.devtech.arrp.json.blockstate.JState;
 import net.devtech.arrp.json.models.JModel;
 import net.devtech.arrp.json.models.JTextures;
-import net.devtech.arrp.json.recipe.JIngredient;
-import net.devtech.arrp.json.recipe.JIngredients;
-import net.devtech.arrp.json.recipe.JRecipe;
-import net.devtech.arrp.json.recipe.JResult;
+import net.devtech.arrp.json.recipe.*;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroup;
-import net.minecraft.item.Items;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 
-public class WarpedBasaltBricksBlock extends Block implements MinekeaBlock {
-    public static final Identifier BLOCK_ID = new Identifier(ModInfo.MOD_ID, "building/warped_basalt_bricks");
+public class WarpedNetherBricksBlock extends Block implements MinekeaBlock {
+    public static final Identifier BLOCK_ID = new Identifier(ModInfo.MOD_ID, "building/warped_nether_bricks");
 
-    public WarpedBasaltBricksBlock() {
-        super(FabricBlockSettings.copyOf(Blocks.SMOOTH_BASALT));
+    public WarpedNetherBricksBlock() {
+        super(FabricBlockSettings.copyOf(Blocks.RED_NETHER_BRICKS));
     }
 
     @Override
@@ -51,18 +48,12 @@ public class WarpedBasaltBricksBlock extends Block implements MinekeaBlock {
 
         MinekeaResourcePack.RESOURCE_PACK.addRecipe(
             BLOCK_ID,
-            JRecipe.shapeless(
-                JIngredients.ingredients()
-                    .add(JIngredient.ingredient().item(BasaltBricksBlock.BLOCK_ID.toString()))
-                    .add(JIngredient.ingredient().item(Items.TWISTING_VINES)),
+            JRecipe.shaped(
+                JPattern.pattern("AB", "BA"),
+                JKeys.keys()
+                    .key("A", JIngredient.ingredient().item(WarpedWartItem.ITEM_ID.toString()))
+                    .key("B", JIngredient.ingredient().item("minecraft:nether_brick")),
                 JResult.result(BLOCK_ID.toString())
-            )
-        );
-        MinekeaResourcePack.RESOURCE_PACK.addRecipe(
-            new Identifier(BLOCK_ID.getNamespace(), BLOCK_ID.getPath() + "_stonecutting_reverse"),
-            JRecipe.stonecutting(
-                JIngredient.ingredient().item(BLOCK_ID.toString()),
-                JResult.stackedResult(BasaltBricksBlock.BLOCK_ID.toString(), 1)
             )
         );
